+++ src/components/SettingsPage/index.tsx
@@ -5,6 +5,7 @@ import { PageContainer } from '../PageContainer';
 import { SaveJSONButton } from './__subComponents/SaveJSONButton';
 import { LoadJSONButton } from './__subComponents/LoadJSONButton';
 import { ReduxState } from '../../modules';
+import { ResetLocalStorageButton } from './__subComponents/ResetLocalStorageButton';
 
 export const SettingsPage: React.FC = () => {
     const version = useSelector((state: ReduxState) => state.app.version);
@@ -13,6 +14,7 @@ export const SettingsPage: React.FC = () => {
             <Typography variant="body1">
                 {`Version: ${version}`}
             </Typography>
+            <ResetLocalStorageButton />
             <SaveJSONButton />
             <LoadJSONButton />
         </PageContainer>